refactor: 매퍼 클래스 레이어 변경
